+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+### v1.20.0 (2023-10-17)
+
 * Add DateIntervalUtils to stringify a DateInterval, and DateIntervalFactory for shorthand creation
 * Add DateTimeImmutableFactory::zeroMicros() to create / modify a DT with microseconds truncated to zero
 * Add DateTimeImmutableFactory::fromIso() to strictly parse ISO 8601 / RFC 3339 date-time strings